+++ lib/honeybadger/cli/main.rb
@@ -6,6 +6,7 @@ require 'honeybadger/cli/notify'
 require 'honeybadger/cli/test'
 require 'honeybadger/config'
 require 'honeybadger/config/defaults'
+require 'honeybadger/ruby'
 require 'honeybadger/util/http'
 require 'honeybadger/version'
 require 'logger'
@@ -137,13 +138,24 @@ WELCOME
       end
 
       def build_config(options)
-        config = Config.new(logger: Logger.new('/dev/null'))
+        load_env
+
+        config = Honeybadger.config
         config.set(:api_key, fetch_value(options, 'api_key')) if options.has_key?('api_key')
         config.set(:env, fetch_value(options, 'environment')) if options.has_key?('environment')
-        config.init!({
-          framework: :cli
-        })
+
         config
+      end
+
+      def load_env
+        # Initialize Rails when running from Rails root.
+        environment_rb = File.join(Dir.pwd, 'config', 'environment.rb')
+        if File.exists?(environment_rb)
+          require environment_rb
+        end
+
+        # Ensure config is loaded (will be skipped if initialized by Rails).
+        Honeybadger.config.load!
       end
 
       def log_error(e)